backend/requirements.txt
flask==3.0.3
sqlalchemy==2.0.32
pg8000==1.31.2
python-dotenv==1.0.1
werkzeug==3.0.3
matplotlib==3.9.2
scikit-learn==1.5.1
flask-sqlalchemy
flask-migrate
psycopg2-binary
python-dotenv

